Add savings goals pages (#3)
* Navigate to show

* Add savings goal `show` page

* Revamp `new` page

* Add `proportion_reached` method

Add a method to savings_goal for calculating the proportion of the goal that has been reached. Subsequently add this rendered value to the show view as a percentage.

* Add savings_goals index view

* Add Shipyard

Add the Shipyard JS framework for styling components. This involved moving application.css to an .scss file in order to leverage the asset pipeline.

* Add table to `show`